backend/requirements.txt
fastapi
uvicorn
python-multipart
elevenlabs
mistralai
google-generativeai
python-dotenv
requests
googlemaps
supabase
python-jose[cryptography]
bcrypt